requirements.txt
absl-py==0.13.0
appdirs==1.4.4
argcomplete==1.12.3
audioread==2.1.9
Bottleneck==1.3.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
ci-sdr==0.0.0
click==8.0.1
ConfigArgParse==1.5.1
configparser==5.0.2
ctc-segmentation==1.6.7
cycler==0.10.0
Cython==0.29.24
decorator==5.0.9
Distance==0.1.3
docker-pycreds==0.4.0
editdistance==0.5.2
einops==0.3.0
espnet==0.10.0
espnet-tts-frontend==0.0.3
filelock==3.0.12
Flask==2.0.1
g2p-en==2.1.0
gdown==3.13.0
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.33.1
google-auth-oauthlib==0.4.4
grpcio==1.39.0
h5py==3.3.0
humanfriendly==9.2
idna==3.2
inflect==5.3.0
itsdangerous==2.0.1
jaconv==0.3
Jinja2==3.0.1
joblib==1.0.1
kaldiio==2.17.2
kiwisolver==1.3.1
librosa==0.8.1
llvmlite==0.36.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.1.0
nara-wpe==0.0.7
nltk==3.6.2
numba==0.53.1
numpy==1.21.1
oauthlib==3.1.1
packaging==21.0
pandas==1.3.1
parallel-wavegan==0.4.8
pathtools==0.1.2
Pillow==8.3.1
pooch==1.4.0
promise==2.3
protobuf==3.17.3
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
pypinyin==0.42.0
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-wpe==0.0.1
pytz==2021.1
pyworld==0.3.0
PyYAML==5.4.1
regex==2021.7.6
requests==2.26.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.0
sentencepiece==0.1.86
sentry-sdk==1.3.0
shortuuid==1.0.1
six==1.16.0
smmap==4.0.0
SoundFile==0.10.3.post1
subprocess32==3.5.4
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
textblob==0.15.3
threadpoolctl==2.2.0
toml==0.10.2
torch-complex==0.2.1
tqdm==4.61.2
typeguard==2.12.1
typing-extensions==3.10.0.0
Unidecode==1.2.0
urllib3==1.26.6
wandb==0.11.0
Werkzeug==2.0.1
xmltodict==0.12.0
yq==2.12.2
